+++ restli-tools/src/main/java/com/linkedin/restli/tools/clientgen/FluentApiGenerator.java
@@ -21,19 +21,16 @@ import com.linkedin.data.schema.resolver.MultiFormatDataSchemaResolver;
 import com.linkedin.internal.tools.ArgumentFileProcessor;
 import com.linkedin.pegasus.generator.CodeUtil;
 import com.linkedin.pegasus.generator.TemplateSpecGenerator;
-import com.linkedin.restli.common.ResourceMethod;
-import com.linkedin.restli.internal.common.RestliVersion;
 import com.linkedin.restli.internal.server.RestLiInternalException;
 import com.linkedin.restli.restspec.ResourceEntityType;
 import com.linkedin.restli.restspec.ResourceSchema;
-import com.linkedin.restli.tools.clientgen.builderspec.BuilderSpec;
+import com.linkedin.restli.tools.clientgen.fluentspec.BaseResourceSpec;
+import com.linkedin.restli.tools.clientgen.fluentspec.CollectionResourceSpec;
+import com.linkedin.restli.tools.clientgen.fluentspec.SpecUtils;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.net.URL;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.GnuParser;
@@ -44,7 +41,6 @@ import org.apache.commons.cli.ParseException;
 import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.VelocityEngine;
 import org.apache.velocity.runtime.RuntimeConstants;
-import org.apache.velocity.runtime.log.Log4JLogChute;
 import org.apache.velocity.runtime.resource.loader.JarResourceLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -59,22 +55,7 @@ public class FluentApiGenerator
 {
   private static final Logger LOGGER = LoggerFactory.getLogger(FluentApiGenerator.class);
   private static final Options OPTIONS = new Options();
-  private static final Map<ResourceMethod, String> BUILDER_BASE_MAP = new HashMap<>();
   private static final String API_TEMPLATE_DIR = "apiVmTemplates";
-
-  static {
-    BUILDER_BASE_MAP.put(ResourceMethod.GET, "get");
-    BUILDER_BASE_MAP.put(ResourceMethod.DELETE, "delete");
-    BUILDER_BASE_MAP.put(ResourceMethod.UPDATE, "update");
-    BUILDER_BASE_MAP.put(ResourceMethod.CREATE, "create");
-    BUILDER_BASE_MAP.put(ResourceMethod.PARTIAL_UPDATE, "partialUpdate");
-    BUILDER_BASE_MAP.put(ResourceMethod.GET_ALL, "getAll");
-    BUILDER_BASE_MAP.put(ResourceMethod.BATCH_GET, "batchGet");
-    BUILDER_BASE_MAP.put(ResourceMethod.BATCH_UPDATE, "batchUpdate");
-    BUILDER_BASE_MAP.put(ResourceMethod.BATCH_PARTIAL_UPDATE, "batchPartialUpdate");
-    BUILDER_BASE_MAP.put(ResourceMethod.BATCH_DELETE, "batchDelete");
-    BUILDER_BASE_MAP.put(ResourceMethod.BATCH_CREATE, "batchCreate");
-  }
 
   public static void main(String[] args) throws Exception
   {
@@ -138,10 +119,9 @@ public class FluentApiGenerator
   {
     final RestSpecParser parser = new RestSpecParser();
     final DataSchemaResolver schemaResolver = MultiFormatDataSchemaResolver.withBuiltinFormats(resolverPath);
+    VelocityEngine velocityEngine = initVelocityEngine();
+    final File targetDirectory = new File(targetDirectoryPath);
 
-
-    final RequestBuilderSpecGenerator specGenerator = new RequestBuilderSpecGenerator(schemaResolver,
-        new TemplateSpecGenerator(schemaResolver), RestliVersion.RESTLI_2_0_0, BUILDER_BASE_MAP);
 
     final RestSpecParser.ParseResult parseResult = parser.parseSources(sources);
 
@@ -155,12 +135,34 @@ public class FluentApiGenerator
         continue;
       }
 
-      try
+      BaseResourceSpec spec = null;
+      if (resourceSchema.hasCollection())
       {
-        specGenerator.generate(resourceSchema, pair.second);
+        spec = new CollectionResourceSpec(resourceSchema,
+            new TemplateSpecGenerator(schemaResolver),
+            pair.second.getPath(),
+            schemaResolver);
+      }
+      else
+      {
+        continue;
+      }
+      File packageDir = new File(targetDirectory, spec.getNamespace().toLowerCase().replace('.', File.separatorChar));
+      packageDir.mkdirs();
+      File file = new File(packageDir, CodeUtil.capitalize(spec.getResource().getName()) + ".java");
+      try (FileWriter writer = new FileWriter(file))
+      {
+        VelocityContext context = new VelocityContext();
+        context.put("spec", spec);
+        context.put("util", SpecUtils.class);
+        if (spec.getResource().hasCollection())
+        {
+          velocityEngine.mergeTemplate(API_TEMPLATE_DIR + "/collection.vm", VelocityEngine.ENCODING_DEFAULT, context, writer);
+        }
       }
       catch (Exception e)
       {
+        LOGGER.error("Error generating fluent client apis", e);
         message.append(e.getMessage()).append("\n");
       }
     }
@@ -170,31 +172,6 @@ public class FluentApiGenerator
       throw new IOException(message.toString());
     }
 
-    Set<BuilderSpec> allBuilders = specGenerator.getBuilderSpec();
-    VelocityEngine velocityEngine = initVelocityEngine();
-    final File targetDirectory = new File(targetDirectoryPath);
-    for (BuilderSpec spec : allBuilders)
-    {
-      File packageDir = new File(targetDirectory, spec.getNamespace().toLowerCase().replace('.', File.separatorChar));
-      packageDir.mkdirs();
-      File file = new File(packageDir, CodeUtil.capitalize(spec.getResource().getName()) + ".java");
-      try (FileWriter writer = new FileWriter(file))
-      {
-        VelocityContext context = new VelocityContext();
-        context.put("spec", spec);
-        context.put("className", CodeUtil.capitalize(spec.getResource().getName()));
-        if (spec.getResource().hasCollection())
-        {
-          // TODO: Implement.
-          velocityEngine.mergeTemplate(API_TEMPLATE_DIR + "/collection.vm", context, writer);
-        }
-      }
-      catch (Exception e)
-      {
-        LOGGER.error("Error generating fluent client apis", e);
-        System.exit(1);
-      }
-    }
   }
 
   private static VelocityEngine initVelocityEngine()
@@ -230,17 +207,14 @@ public class FluentApiGenerator
       velocity = new VelocityEngine();
 
       final String resourceDirPath = new File(templateDirUrl.getPath()).getParent();
-      configName = new StringBuilder("file.").append(VelocityEngine.RESOURCE_LOADER).append(".path");
-      velocity.setProperty(configName.toString(), resourceDirPath);
+      velocity.setProperty(RuntimeConstants.FILE_RESOURCE_LOADER_PATH, resourceDirPath);
     }
     else
     {
       throw new IllegalArgumentException("Unsupported template path scheme");
     }
-
-    velocity.setProperty(RuntimeConstants.RUNTIME_LOG_LOGSYSTEM_CLASS, Log4JLogChute.class.getName());
-    velocity.setProperty(Log4JLogChute.RUNTIME_LOG_LOG4J_LOGGER, FluentApiGenerator.class.getName());
-
+    velocity.setProperty(RuntimeConstants.SPACE_GOBBLING, RuntimeConstants.SpaceGobbling.STRUCTURED.name());
+    velocity.setProperty(RuntimeConstants.VM_LIBRARY, "macros/library.vm");
     try
     {
       velocity.init();